FIX: 통합 예외 처리
