+++ src/main/java/com/tidavid1/Studywith/global/common/exception/ErrorCode.java
@@ -1,0 +1,12 @@
+package com.tidavid1.Studywith.global.common.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum ErrorCode {
+    UNKNOWN(-9999, "알 수 없는 오류");
+    private final int code;
+    private final String message;
+}
