+++ src/main/java/com/tidavid1/Studywith/global/common/exception/GlobalExceptionHandler.java
@@ -1,0 +1,19 @@
+package com.tidavid1.Studywith.global.common.exception;
+
+import com.tidavid1.Studywith.global.common.response.model.CommonResult;
+import com.tidavid1.Studywith.global.common.response.service.ResponseFactory;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+@RequiredArgsConstructor
+public class GlobalExceptionHandler {
+    @ExceptionHandler(Exception.class)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    protected CommonResult defaultException(Exception e){
+        return ResponseFactory.getFailResult(ErrorCode.UNKNOWN.getCode(), ErrorCode.UNKNOWN.getMessage());
+    }
+}
