--- src/main/java/com/tidavid1/Studywith/global/exception/GlobalExceptionHandler.java
@@ -1,34 +1,0 @@
-package com.tidavid1.Studywith.global.exception;
-
-import com.tidavid1.Studywith.global.common.response.model.CommonResult;
-import com.tidavid1.Studywith.global.common.response.service.ResponseService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.MessageSource;
-import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-@RequiredArgsConstructor
-public class GlobalExceptionHandler {
-    private final ResponseService responseService;
-    private final MessageSource messageSource;
-
-    private String getMessage(String code){
-        return getMessage(code, null);
-    }
-
-    private String getMessage(String code, Object[] args){
-        return messageSource.getMessage(code, args, LocaleContextHolder.getLocale());
-    }
-
-    @ExceptionHandler(Exception.class)
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    protected CommonResult defaultException(Exception e){
-        return responseService.getFailResult(
-                999, e.getMessage()
-        );
-    }
-}
