+++ src/test/java/com/tidavid1/Studywith/global/common/exception/GlobalExceptionHandlerTest.java
@@ -1,34 +1,27 @@
-package com.tidavid1.Studywith.global.exception;
+package com.tidavid1.Studywith.global.common.exception;
 
 import com.tidavid1.Studywith.global.common.response.model.CommonResult;
-import com.tidavid1.Studywith.global.common.response.service.ResponseService;
+import com.tidavid1.Studywith.global.common.response.service.ResponseFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
+
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.MessageSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.doThrow;
+
 
 @SpringBootTest
 @ExtendWith(MockitoExtension.class)
 class GlobalExceptionHandlerTest {
 
-    ResponseService responseService;
-    @Mock
-    MessageSource messageSource;
-    @InjectMocks
     GlobalExceptionHandler globalExceptionHandler;
 
     @BeforeEach
     void setup(){
-        responseService = new ResponseService();
-        globalExceptionHandler = new GlobalExceptionHandler(responseService, messageSource);
+        globalExceptionHandler = new GlobalExceptionHandler();
     }
     @DisplayName("Test DefaultException")
     @Test
@@ -36,7 +29,8 @@ class GlobalExceptionHandlerTest {
         // Arrange
         Exception exception = new Exception();
         CommonResult actualResult = new CommonResult();
-        actualResult.setCode(999);
+        actualResult.setCode(ErrorCode.UNKNOWN.getCode());
+        actualResult.setMessage(ErrorCode.UNKNOWN.getMessage());
         // Act
         CommonResult expectedResult = globalExceptionHandler.defaultException(exception);
         // Assert
